Lots of refactoring and renaming
Please note that capitalisation has changed for a handful of public-facing members in order to ensure consistency. Given how moonsharp's interpreter works, existing scripts MAY (but also may not) need to be updated.

I did most of the actual work here months ago and just never committed it, so this one's a mess and that's on me.