Add map flags, abstract world positioning
It's now possible to set the map flag marker from scripts, and while it may not LOOK like it's cleaner, world positioning is cleaned up a little by abstracting common parts into a new WorldPosition class. The new IWorldObjectWrapper interface is designed to expose world/map positioning for things that exist in the game world, so there don't need to be separate overloads for every single type.